github-maintainer-ai/requirements.txt
PyGithub>=2.1.1
python-dotenv>=1.0.0
requests>=2.31.0
loguru>=0.7.2
rich>=13.6.0
langchain>=0.0.330
ollama>=0.1.0
pydantic>=2.4.2

